feat(docker): skip installing docker if docker command exists

+++ docker/install.sh
@@ -2,6 +2,9 @@
 source $(dirname $(dirname ${BASH_SOURCE}))/framework/oo-bootstrap.sh
 
 if [[ $(OS::LSBDist) == "centos" ]]; then
+    if Command::Exists docker; then
+        exit 0
+    fi
     sudo yum install -y yum-utils device-mapper-persistent-data lvm2
     sudo yum-config-manager --add-repo https://download.docker.com/linux/centos/docker-ce.repo
     sudo yum install -y docker-ce
